+++ examples/regression_YearPredictionMSD_mlp.py
@@ -13,15 +13,17 @@ from torchensemble.fusion import FusionRegressor
 from torchensemble.voting import VotingRegressor
 from torchensemble.bagging import BaggingRegressor
 from torchensemble.gradient_boosting import GradientBoostingRegressor
+from torchensemble.utils import set_logger
 
 
-def load_data(batch_size):
+def load_data(batch_size, logger):
 
     # The dataset can be downloaded from:
     #   https://www.csie.ntu.edu.tw/~cjlin/libsvmtools/datasets/regression.html#YearPredictionMSD
 
     if not isinstance(batch_size, numbers.Integral):
         msg = "`batch_size` should be an integer, but got {} instead."
+        logger.error(msg.format(batch_size))
         raise ValueError(msg.format(batch_size))
 
     # MODIFY THE PATH IF YOU WANT
@@ -52,7 +54,7 @@ def load_data(batch_size):
     return train_loader, test_loader
 
 
-def display_records(records):
+def display_records(records, logger):
     msg = (
         "{:<28} | Testing MSE: {:.2f} | Training Time: {:.2f} s |"
         " Evaluating Time: {:.2f} s"
@@ -60,7 +62,7 @@ def display_records(records):
 
     print("\n")
     for method, training_time, evaluating_time, mse in records:
-        print(msg.format(method, mse, training_time, evaluating_time))
+        logger.info(msg.format(method, mse, training_time, evaluating_time))
 
 
 class MLP(nn.Module):
@@ -96,16 +98,19 @@ if __name__ == "__main__":
     records = []
     torch.manual_seed(0)
 
+    logger = set_logger("INFO", "regression_YearPredictionMSD_mlp", "DEBUG")
+
     # Load data
-    train_loader, test_loader = load_data(batch_size)
-    print("Finish loading data...\n")
+    train_loader, test_loader = load_data(batch_size, logger)
+    # print("Finish loading data...\n")
+    logger.info("Finish loading data...\n")
 
     # FusionRegressor
     model = FusionRegressor(
         estimator=MLP,
         n_estimators=n_estimators,
         cuda=True,
-        n_jobs=1,
+        n_jobs=1
     )
 
     tic = time.time()
@@ -126,7 +131,7 @@ if __name__ == "__main__":
         estimator=MLP,
         n_estimators=n_estimators,
         cuda=True,
-        n_jobs=1,
+        n_jobs=1
     )
 
     tic = time.time()
@@ -147,7 +152,7 @@ if __name__ == "__main__":
         estimator=MLP,
         n_estimators=n_estimators,
         cuda=True,
-        n_jobs=1,
+        n_jobs=1
     )
 
     tic = time.time()
@@ -184,4 +189,4 @@ if __name__ == "__main__":
                     evaluating_time, testing_mse))
 
     # Print results on different ensemble methods
-    display_records(records)
+    display_records(records, logger)